finished lesson 3

+++ src/com/team2052/KnightKrawlerIntroJava/Lesson03/MultiCalculator.java
@@ -3,28 +3,28 @@ package com.team2052.KnightKrawlerIntroJava.Lesson03;
 import java.util.Scanner;
 
 public class MultiCalculator {
-    private int number1;
-    private  int number2;
+    private double number1;
+    private double number2;
     private Operation calcOperation = Operation.ADD;
 
 
-        public int getNumber1() {
+        public double getNumber1() {
             return number1;
         }
 
-        public void setNumber1(int val) {
+        public void setNumber1(double val) {
             number1 = val;
         }
 
-        public int getNumber2() {
+        public double getNumber2() {
             return number2;
         }
 
-        public void setNumber2(int val) {
+        public void setNumber2(double val) {
             number2 = val;
         }
 
-        public int calculate() {
+        public double calculate() {
             if (getCalcOperation() == Operation.ADD) {
                 return this.getNumber1() + this.getNumber2();
